clippy: apply auto-fix safe lints (1020 → 339 warnings)
Run `cargo clippy --release --all-targets --fix --allow-dirty` to apply
all auto-fixable lint suggestions across the workspace. Net change:
1020 warnings → 339 (-681). All tests still pass:

  mdpicoem-common  234 / 235 ok (1 ignored, gated)
  mdpicoem-devices  35 / 35 ok
  mdrp2040         850 + 24 ok (lib + integration)
  mdrp2350        2719 + 21 ok (lib + integration; 30 ignored, gated)

Categories cleaned up by --fix:
  identity_op, single_match, unnecessary_cast, manual_range_contains,
  manual_div_ceil, manual_is_multiple_of, redundant_field_names,
  collapsible_if, useless_vec, needless_range_loop, etc.

Categories left for follow-up (see the next commit):
  doc_list_item_without_indentation (60), inconsistent_digit_grouping
  (55), field_reassign_with_default (35), assertions_on_constants (20),
  push_after_create (12), too_many_arguments (8+4) — design feedback,
  upper_case_acronyms (HANDLE/DWORD/BOOL — winapi-derived names),
  needless_range_loop (idx-into-array test patterns).

+++ crates/mdpicoem-harness/src/cycle_cases.rs
@@ -1004,7 +1004,7 @@ mod tests {
         let instr_addr_hw = 19;
         let pc_hw = instr_addr_hw + 2; // PC = instr+4 bytes = +2 halfwords
         let target_hw: i32 = 2; // poll label
-        let imm11_hw = target_hw - pc_hw as i32; // signed halfword offset
+        let imm11_hw = target_hw - pc_hw; // signed halfword offset
         let encoded = 0xE000u16 | ((imm11_hw as i16) as u16 & 0x07FF);
         assert_eq!(
             MEASUREMENT_STUB[19], encoded,
